Fix quiz attempt
restore header and implement the feedback on screen for the score, make the timer only stop running once the quiz is fully finished, and make it so the user can continue writing the quiz until all the questions are answered, even if submit is hit before it's finished. All it will do is remove the ability to change answered questions.